Worker/requirements.txt
-i https://pypi.org/simple
boto3==1.9.128
botocore==1.12.128
docutils==0.14
jmespath==0.9.4
pymongo==3.7.2
python-dateutil==2.8.0 ; python_version >= '2.7'
python-decouple==3.1
s3transfer==0.2.0
six==1.12.0
urllib3==1.24.1 ; python_version >= '3.4'
